+++ plugins/dss/hooks.ts
@@ -1,7 +1,5 @@
 import { MountService } from "foundation-ui";
 
-import VolumeMountTypesSelect from "./components/VolumeMountTypesSelect";
-import VolumeTypeSelect from "./components/VolumeTypeSelect";
 import DSSInput from "./components/DSSInput";
 import DSSVolumeConfig from "./components/DSSVolumeConfig";
 
@@ -11,18 +9,9 @@ module.exports = {
       DSSVolumeConfig,
       "CreateService:SingleContainerVolumes:UnknownVolumes"
     );
-
     MountService.MountService.registerComponent(
       DSSInput,
       "CreateService:MultiContainerVolumes:UnknownVolumes"
-    );
-    MountService.MountService.registerComponent(
-      VolumeMountTypesSelect,
-      "CreateService:MultiContainerVolumes:Types"
-    );
-    MountService.MountService.registerComponent(
-      VolumeTypeSelect,
-      "CreateService:SingleContainerVolumes:Types"
     );
   },
 };
